pyproject.toml
[project]
name = "rnacentral_pipeline"
version = "0.1.0"
description = "The pipeline that imports all RNAcentral data"
authors = ["Blake Sweeney <bsweeney@ebi.ac.uk>"]
package-mode = false
requires-python = ">=3.11"

dependencies = [
    "PyMySQL>=1.0.2",
    "attrs>=21.4.0",
    "beautifulsoup4>=4.11.0",
    "biopython>=1.81",
    "click>=8.1.3",
    "click-aliases>=1.0.1",
    "furl>=2.1.3",
    "gffutils>=0.11.1",
    "humanfriendly>=10.0",
    "ijson>=3.1.4",
    "intervaltree>=3.1.0",
    "jsonschema>=4.17",
    "lxml>=4.9.1",
    "more-itertools>=9.0",
    "obonet==0.3.0",
    "pandas>=1.5.1",
    "PyPika>=0.48.8",
    "pyyaml>=6.0",
    "requests>=2.28.1",
    "retry>=0.9.2",
    "scikit-learn==1.2.2",
    "semver>=2.13.0",
    "setuptools",
    "slack_sdk>=3.19.4",
    "sqlitedict>=2.0.0",
    "TatSu==5.10.6",
    "throttler>=1.2.0",
    "nltk>=3.8.1",
    "openpyxl>=3.0.10",
    "xlrd>=2.0.1",
    "pybedtools>=0.12.0",
    "psycopg2-binary>=2.9.7",
    "polars>=1.14.0",
    "numpy<2",
    "gensim>=4.3.3",
    "python-louvain>=0.16",
    "pysam>=0.23.3",
    "onnxruntime>=1.22.0",
    "cython>=3.1.2",
    "skl2onnx>=1.19.1",
    "datasets>=3.5.0",
    "fastjsonschema>=2.21.2",
]


[dependency-groups]
dev = [
    "black>=24.4.2",
    "isort>=5.13.2",
    "pytest>=7.2.0",
    "maturin>=1.10",
    "line-profiler>=5.0.0",
]

[tool.pytest.ini_options]
xfail_strict = true
filterwarnings = "ignore::DeprecationWarning"
markers = [
    "slow: Tests that take a long time",
    "db: Test that require access to our database",
]
